+++ src/context/Context.ts
@@ -1,0 +1,6 @@
+import { createContext } from "react";
+import { ContextType } from "../types";
+
+const Context = createContext({} as ContextType);
+
+export default Context;
